+++ source/tryGoal/component/tryGoalController.js
@@ -42,6 +42,7 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
     $scope.runGoal = function(goalInfo) {
         var requiresInitData = false;
         $scope.init = {};
+        $scope.response = {};
         $rootScope.apiOutput = new Array();
         sessionId = null;
         contextId = goalInfo.contextId;
@@ -111,6 +112,31 @@ function($scope, agentMemory, $compile, $stateParams, config, GoalAPI, ConfigAPI
 
     $scope.toggleSplitScreen = function() {
         $scope.$parent.splitScreen = !$scope.$parent.splitScreen;
+    };
+
+    $scope.toggleQueryContext = function () {
+        $scope.$parent.displayQueryContext = !$scope.$parent.displayQueryContext;
+    }
+
+    function formatConcept(concept) {
+        if (concept === '(null)') {
+            return '(*)';
+        } else {
+            return concept;
+        }
+    };
+
+    $scope.formatRule = function (rule) {
+        var concepts = rule.match(/\((.*?)\)/gm);
+        var relationship = rule.match(/\[(.*?)\]/);
+
+        if (concepts && relationship) {
+            concepts[0] = formatConcept(concepts[0]);
+            concepts[1] = formatConcept(concepts[1]);
+            return concepts[0] + ' ' + relationship[1] + ' ' + concepts[1] + ' ' + concepts[2];
+        } else {
+            return 'There has been a problem presenting this rule.';
+        }
     };
 
     $scope.queryGoal = function() {