+++ CHANGELOG.md
@@ -11,6 +11,7 @@ facing release notes.
   * [Misc] Other items
 
 ## v2.29.0
+  * [New]   RB-3552: Add button to Try Query to show context of question.
   * [New]   RB-3668: Add timeout to pdf download button, and condition to button.
   * [New]   RB-3558: Enable PDF downloads of agent interactions